move utils to uplc and break up
