+++ crates/uplc/src/transaction_eval/script_context.rs
@@ -1,0 +1,85 @@
+use pallas_codec::utils::KeyValuePairs;
+use pallas_crypto::hash::Hash;
+use pallas_primitives::babbage::{
+    AddrKeyhash, Certificate, Coin, DatumHash, Mint, PlutusData, PolicyId, Redeemer, RewardAccount,
+    StakeCredential, TransactionInput, TransactionOutput, Value, Withdrawals,
+};
+
+#[derive(Debug, PartialEq, Clone)]
+pub struct ResolvedInput {
+    pub input: TransactionInput,
+    pub output: TransactionOutput,
+}
+
+#[derive(Debug, PartialEq, Clone)]
+pub struct TxInInfo {
+    pub out_ref: TransactionInput,
+    pub resolved: TxOut,
+}
+#[derive(Debug, PartialEq, Clone)]
+pub enum TxOut {
+    V1(TransactionOutput),
+    V2(TransactionOutput),
+}
+
+#[derive(Debug, PartialEq, Eq, Clone)]
+pub enum ScriptPurpose {
+    Minting(PolicyId),
+    Spending(TransactionInput),
+    Rewarding(StakeCredential),
+    Certifying(Certificate),
+}
+
+#[derive(Debug, PartialEq, Clone)]
+pub struct TxInfoV1 {
+    pub inputs: Vec<TxInInfo>,
+    pub outputs: Vec<TxOut>,
+    pub fee: Value,
+    pub mint: Mint,
+    pub dcert: Vec<Certificate>,
+    pub wdrl: Vec<(RewardAccount, Coin)>,
+    pub valid_range: TimeRange,
+    pub signatories: Vec<AddrKeyhash>,
+    pub data: Vec<(DatumHash, PlutusData)>,
+    pub id: Hash<32>,
+}
+
+#[derive(Debug, PartialEq, Clone)]
+pub struct TxInfoV2 {
+    pub inputs: Vec<TxInInfo>,
+    pub reference_inputs: Vec<TxInInfo>,
+    pub outputs: Vec<TxOut>,
+    pub fee: Value,
+    pub mint: Mint,
+    pub dcert: Vec<Certificate>,
+    pub wdrl: Withdrawals,
+    pub valid_range: TimeRange,
+    pub signatories: Vec<AddrKeyhash>,
+    pub redeemers: KeyValuePairs<ScriptPurpose, Redeemer>,
+    pub data: KeyValuePairs<DatumHash, PlutusData>,
+    pub id: Hash<32>,
+}
+
+#[derive(Debug, PartialEq, Clone)]
+pub enum TxInfo {
+    V1(TxInfoV1),
+    V2(TxInfoV2),
+}
+
+#[derive(Debug, PartialEq, Clone)]
+pub struct ScriptContext {
+    pub tx_info: TxInfo,
+    pub purpose: ScriptPurpose,
+}
+
+//---- Time conversion: slot range => posix time range
+#[derive(Debug, PartialEq, Clone)]
+pub struct TimeRange {
+    pub lower_bound: Option<u64>,
+    pub upper_bound: Option<u64>,
+}
+
+pub struct SlotConfig {
+    pub slot_length: u64,
+    pub zero_time: u64,
+}
